+++ cmd/influxd/handler.go
@@ -2,6 +2,7 @@ package main
 
 import (
 	"log"
+	"math/rand"
 	"net/http"
 	"net/url"
 	"strings"
@@ -73,7 +74,7 @@ func (h *Handler) serveMessaging(w http.ResponseWriter, r *http.Request) {
 // serveMetadata responds to broker requests
 func (h *Handler) serveMetadata(w http.ResponseWriter, r *http.Request) {
 	if h.Broker == nil && h.Server == nil {
-		log.Println("no broker or server configured to handle messaging endpoints")
+		log.Println("no broker or server configured to handle metadata endpoints")
 		http.Error(w, "server unavailable", http.StatusServiceUnavailable)
 		return
 	}
@@ -98,7 +99,7 @@ func (h *Handler) serveMetadata(w http.ResponseWriter, r *http.Request) {
 // serveRaft responds to raft requests.
 func (h *Handler) serveRaft(w http.ResponseWriter, r *http.Request) {
 	if h.Log == nil && h.Server == nil {
-		log.Println("no broker or server configured to handle messaging endpoints")
+		log.Println("no broker or server configured to handle raft endpoints")
 		http.Error(w, "server unavailable", http.StatusServiceUnavailable)
 		return
 	}
@@ -116,7 +117,7 @@ func (h *Handler) serveRaft(w http.ResponseWriter, r *http.Request) {
 // serveData responds to data requests
 func (h *Handler) serveData(w http.ResponseWriter, r *http.Request) {
 	if h.Broker == nil && h.Server == nil {
-		log.Println("no broker or server configured to handle messaging endpoints")
+		log.Println("no broker or server configured to handle data endpoints")
 		http.Error(w, "server unavailable", http.StatusServiceUnavailable)
 		return
 	}
@@ -151,5 +152,5 @@ func (h *Handler) redirect(u []url.URL, w http.ResponseWriter, r *http.Request) 
 	// this is happening frequently, the clients are using a suboptimal endpoint
 
 	// Redirect the client to a valid data node that can handle the request
-	http.Redirect(w, r, u[0].String()+r.RequestURI, http.StatusTemporaryRedirect)
+	http.Redirect(w, r, u[rand.Intn(len(u))].String()+r.RequestURI, http.StatusTemporaryRedirect)
 }